+++ organize/core.py
@@ -1,10 +1,11 @@
 import logging
 import os
-from collections import Counter
+from collections import Counter, defaultdict
 from datetime import datetime
 from typing import Iterable, NamedTuple
 
 import fs
+from fs.errors import NoSysPath
 from fs.base import FS
 from fs.walk import Walker
 from rich.console import Console
@@ -16,7 +17,14 @@ from .actions.action import Action
 from .config import CONFIG_SCHEMA, load_from_file
 from .filters import FILTERS
 from .filters.filter import Filter
-from .utils import Template, deep_merge_inplace, ensure_list, ensure_dict, to_args
+from .utils import (
+    Template,
+    deep_merge_inplace,
+    ensure_list,
+    ensure_dict,
+    to_args,
+    flatten_all_lists_in_dict,
+)
 
 logger = logging.getLogger(__name__)
 highlighted_console = Console()
@@ -40,6 +48,19 @@ DEFAULT_SYSTEM_EXCLUDE_DIRS = [
     ".git",
     ".svn",
 ]
+
+
+def config_cleanup(rules):
+    result = defaultdict(list)
+
+    # delete every root key except "rules"
+    for rule in rules.get("rules", []):
+        # delete disabled rules
+        if rule.get("enabled", True):
+            result["rules"].append(rule)
+
+    # flatten all lists everywhere
+    return flatten_all_lists_in_dict(dict(result))
 
 
 def walker_args_from_location_options(options):
@@ -110,11 +131,15 @@ def instantiate_action(action_config):
     return ACTIONS[name](*args, **kwargs)
 
 
+def syspath_or_exception(fs, path):
+    try:
+        return fs.getsyspath(path)
+    except NoSysPath as e:
+        return e
+
+
 def replace_with_instances(config):
     warnings = []
-
-    # delete disabled rules
-    config["rules"] = [rule for rule in config["rules"] if rule.get("enabled", True)]
 
     for rule in config["rules"]:
         _locations = []
@@ -153,28 +178,39 @@ def replace_with_instances(config):
     return warnings
 
 
-def filter_pipeline(filters: Iterable[Filter], args: dict) -> bool:
+def filter_pipeline(filters: Iterable[Filter], args: dict, filter_mode: str) -> bool:
     """
     run the filter pipeline.
     Returns True on a match, False otherwise and updates `args` in the process.
     """
+    results = []
     for filter_ in filters:
         try:
             match, updates = filter_.pipeline(args)
-            if not (match ^ filter_.inverted):
+            result = match ^ filter_.inverted
+            # we cannot exit early on "any".
+            if (filter_mode == "none" and result) or (
+                filter_mode == "all" and not result
+            ):
                 return False
+            results.append(result)
             deep_merge_inplace(args, updates)
         except Exception as e:  # pylint: disable=broad-except
             logger.exception(e)
             # console.print_exception()
             filter_.print_error(str(e))
             return False
+
+    if filter_mode == "any":
+        return any(results)
     return True
 
 
 def action_pipeline(actions: Iterable[Action], args: dict, simulate: bool) -> bool:
     for action in actions:
         try:
+            # update path
+            args["path"] = syspath_or_exception(args["fs"], args["fs_path"])
             updates = action.pipeline(args, simulate=simulate)
             # jobs may return a dict with updates that should be merged into args
             if updates is not None:
@@ -196,6 +232,7 @@ def run(config, simulate: bool = True):
     for rule_nr, rule in enumerate(config["rules"], start=1):
         target = rule.get("targets", "files")
         console.rule(rule.get("name", "Rule %s" % rule_nr))
+        filter_mode = rule.get("filter_mode", "all")
 
         for walker, base_fs, base_path in rule["locations"]:
             console.location(base_fs, base_path)
@@ -210,11 +247,12 @@ def run(config, simulate: bool = True):
                     "env": os.environ,
                     "now": datetime.now(),
                     "utcnow": datetime.utcnow(),
-                    "path": lambda: base_fs.getsyspath(path),
+                    "path": syspath_or_exception(base_fs, path),
                 }
                 match = filter_pipeline(
                     filters=rule["filters"],
                     args=args,
+                    filter_mode=filter_mode,
                 )
                 if match:
                     is_success = action_pipeline(
@@ -237,6 +275,7 @@ def run_file(config_file: str, working_dir: str, simulate: bool):
     console.info(config_file, working_dir)
     try:
         rules = load_from_file(config_file)
+        rules = config_cleanup(rules)
         CONFIG_SCHEMA.validate(rules)
         warnings = replace_with_instances(rules)
         for msg in warnings:
@@ -245,7 +284,9 @@ def run_file(config_file: str, working_dir: str, simulate: bool):
         count = run(rules, simulate=simulate)
         console.summary(count)
     except SchemaError as e:
-        console.error("Invalid config file")
-        highlighted_console.print(e.autos[-1])
+        console.error("Invalid config file!")
+        for err in e.autos:
+            if err:
+                highlighted_console.print(err)
     except Exception as e:
         highlighted_console.print_exception()